+++ src/Exceptions/UnknownEnumClass.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Kwaadpepper\Enum\Exceptions;
+
+use Exception;
+
+class UnknownEnumClass extends Exception
+{
+    //
+}