+++ changelog.md
@@ -9,6 +9,7 @@ All notable changes to `Enum` will be documented in this file.
 ## Version 0.0.2
 
 ### - Add
+
 - Enum cast in Models
 
     usage:
@@ -23,6 +24,7 @@ All notable changes to `Enum` will be documented in this file.
 ## Version 0.1.0
 
 ### - Changed
+
 - toArray Method now returns an array of BaseEnums
   this is because php numeric string array indexes to int, this would break
   origin enum values
@@ -30,9 +32,22 @@ All notable changes to `Enum` will be documented in this file.
 ## Version 0.1.1
 
 ### Fixed
-  - fixed performances issue on toValues and toLabels
+
+- Fixed performances issue on toValues and toLabels
 
 ## Version 0.1.2
 
 ### Changed
-  - changed toJson to output array with label and value
+
+- Changed toJson to output array with label and value
+
+## Version 0.1.3
+
+### Added
+
+- Validation rule EnumIsValidRule, the constructor needs a BaseEnum class
+  eg: new EnumIsValidRule(ArcadiaMailType::class)
+
+### Fixed
+
+- Unit tests with 0.1.2 changes